+++ src/System/Logger/Settings.hs
@@ -22,7 +22,11 @@ module System.Logger.Settings
     , netstrings
     , setNetStrings
     , logLevel
+    , logLevelMap
+    , logLevelOf
     , setLogLevel
+    , setLogLevelMap
+    , setLogLevelOf
     , name
     , setName
     , nameMsg
@@ -32,19 +36,21 @@ module System.Logger.Settings
 import Data.String
 import Data.ByteString (ByteString)
 import Data.ByteString.Char8 (pack)
+import Data.Map.Strict as Map
 import Data.Text (Text)
 import System.Log.FastLogger (defaultBufSize)
 import System.Logger.Message
 
 data Settings = Settings
-    { _logLevel   :: !Level      -- ^ messages below this log level will be suppressed
-    , _output     :: !Output     -- ^ log sink
-    , _format     :: !DateFormat -- ^ the timestamp format (use \"\" to disable timestamps)
-    , _delimiter  :: !ByteString -- ^ text to intersperse between fields of a log line
-    , _netstrings :: !Bool       -- ^ use <http://cr.yp.to/proto/netstrings.txt netstrings> encoding (fixes delimiter to \",\")
-    , _bufSize    :: !Int        -- ^ how many bytes to buffer before commiting to sink
-    , _name       :: !Text       -- ^ logger name
-    , _nameMsg    :: Msg -> Msg
+    { _logLevel   :: !Level            -- ^ messages below this log level will be suppressed
+    , _levelMap   :: !(Map Text Level) -- ^ log level per named logger
+    , _output     :: !Output           -- ^ log sink
+    , _format     :: !DateFormat       -- ^ the timestamp format (use \"\" to disable timestamps)
+    , _delimiter  :: !ByteString       -- ^ text to intersperse between fields of a log line
+    , _netstrings :: !Bool             -- ^ use <http://cr.yp.to/proto/netstrings.txt netstrings> encoding (fixes delimiter to \",\")
+    , _bufSize    :: !Int              -- ^ how many bytes to buffer before commiting to sink
+    , _name       :: !(Maybe Text)     -- ^ logger name
+    , _nameMsg    :: !(Msg -> Msg)
     }
 
 output :: Settings -> Output
@@ -53,6 +59,7 @@ output = _output
 setOutput :: Output -> Settings -> Settings
 setOutput x s = s { _output = x }
 
+-- | The time and date format used for the timestamp part of a log line.
 format :: Settings -> DateFormat
 format = _format
 
@@ -65,12 +72,15 @@ bufSize = _bufSize
 setBufSize :: Int -> Settings -> Settings
 setBufSize x s = s { _bufSize = max 1 x }
 
+-- | Delimiter string which separates log line parts.
 delimiter :: Settings -> ByteString
 delimiter = _delimiter
 
 setDelimiter :: ByteString -> Settings -> Settings
 setDelimiter x s = s { _delimiter = x }
 
+-- | Whether to use <http://cr.yp.to/proto/netstrings.txt netstring>
+-- encoding for log lines.
 netstrings :: Settings -> Bool
 netstrings = _netstrings
 
@@ -83,12 +93,28 @@ logLevel = _logLevel
 setLogLevel :: Level -> Settings -> Settings
 setLogLevel x s = s { _logLevel = x }
 
-name :: Settings -> Text
+-- | Log level of some named logger.
+logLevelOf :: Text -> Settings -> Maybe Level
+logLevelOf x s = Map.lookup x (_levelMap s)
+
+logLevelMap :: Settings -> Map Text Level
+logLevelMap = _levelMap
+
+-- | Specify a log level for the given named logger. When a logger is
+-- 'clone'd and given a name, the 'logLevel' of the cloned logger will be
+-- the provided here.
+setLogLevelOf :: Text -> Level -> Settings -> Settings
+setLogLevelOf n x s = s { _levelMap = Map.insert n x (_levelMap s) }
+
+setLogLevelMap :: Map Text Level -> Settings -> Settings
+setLogLevelMap x s = s { _levelMap = x }
+
+name :: Settings -> Maybe Text
 name = _name
 
-setName :: Text -> Settings -> Settings
-setName "" s = s { _name = "", _nameMsg = id }
-setName xs s = s { _name = xs, _nameMsg = "logger" .= xs }
+setName :: Maybe Text -> Settings -> Settings
+setName Nothing   s = s { _name = Nothing, _nameMsg = id }
+setName (Just xs) s = s { _name = Just xs, _nameMsg = "logger" .= xs }
 
 nameMsg :: Settings -> (Msg -> Msg)
 nameMsg = _nameMsg
@@ -133,8 +159,8 @@ iso8601UTC = "%Y-%0m-%0dT%0H:%0M:%0SZ"
 --
 --   * 'bufSize'    = 'FL.defaultBufSize'
 --
---   * 'name'       = \"\"
+--   * 'name'       = Nothing
 --
 defSettings :: Settings
-defSettings = Settings Debug StdOut iso8601UTC ", " False defaultBufSize "" id
+defSettings = Settings Debug Map.empty StdOut iso8601UTC ", " False defaultBufSize Nothing id