+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.12.0
+-----------------------------------------------------------------------------
+- Add support for logger-specific log-levels.
+- Changed logger `name` in settings to `Maybe Text`.
+
 0.10
 -----------------------------------------------------------------------------
 - Introduce `Settings` module.